+++ java/src/tarehart/rlbot/CarRotation.java
@@ -2,13 +2,13 @@ package tarehart.rlbot;
 
 public class CarRotation {
 
-    public float noseX;
-    public float noseY;
-    public float noseZ;
-    public float roofZ;
-    public float rollRightDesire;
+    public double noseX;
+    public double noseY;
+    public double noseZ;
+    public double roofZ;
+    public double rollRightDesire;
 
-    public CarRotation(float noseX, float noseY, float noseZ, float roofZ, float rollRightDesire) {
+    public CarRotation(double noseX, double noseY, double noseZ, double roofZ, double rollRightDesire) {
         this.noseX = noseX;
         this.noseY = noseY;
         this.noseZ = noseZ;
